refactor: Split ledger test suite into two files
